fix iopoll probing on system that return success for 0-sized reads
In a perfect world, we would issue a read of size 0 because we are not
interested in any data, just to check whether we _could_. Unfortunately,
it seems we are not allowed to have nice things, and the kernel can
short-circuit its internal logic and return an early "success" for reads
of size 0. We are therefore forced to do an actual read.