fix the conversion from `Duration` to `TimeSpec64`
`std::time::Duration` uses `u64` to represent seconds while the kernel
uses signed integers. Our implementation of `From` would cast directly,
resulting in two's complement wrapping.

This commit removes `From` in favor of `TryFrom`. This new
implementation allows the user to know whether the returned `TimeSpec64`
maps the original duration or if it was capped to the maximum
value a `TimeSpec64` can represent.

It also makes it easy for a user to ignore the saturation (presumably
this will only ever trigger when trying to convert `Duration::MAX`
values and the likes).

Co-authored-by: Shamir Khodzha <khodzha.sh@gmail.com>
Co-authored-by: Hippolyte Barraud <hippolyte.barraud@datadoghq.com>