sleep by `poll`ing an eventfd
We no longer link the rings at all. Instead, we register an eventfd on
the rings (that is written to by the kernel each time a CQE is posted)
and we poll on that instead.

It simplifies the sleeping procedure quite a bit at the expense of
adding one (nonblocking) `read` syscall per reactor trip.