Fix `pre_allocate` implementation
Previously invoking `pre_allocate` would zero all data in the range [0,
size). Change it so that it follows the expected behavior of how
fallocate documents file ranges are reserved so that existing data isn't
modified. Additionally add a new option to be able to specify whether or
not you want the reported size of the file to change because of this
pre-allocation.

Fixes #576.