Fix background thread pool unbound placement
Any threads we spawn inherit our affinity. This means that background
thread pools for any executors that don't have an affinity of "Unbound"
will only ever run on the executor's CPU rather than distributing work
across many CPUs. This seems undesirable given that the background
thread pool placement can be set independent of the executor it's
associated with. Additionally, the background thread pool placement by
default uses the pool placement so this has no impact on background
threads that didn't have a custom placement set (& if the placement was
set, it was likely the wrong thing happening).