Fix Unbound background thread pool affinity (#607)
* fix CI: Bump minimum version to 1.65

backtrace picked up a new addr2line in a patch release that depends on
1.65.

* Fix Cargo warning about resolver

```
warning: some crates are on edition 2021 which defaults to `resolver = "2"`, but virtual workspaces default to `resolver = "1"`
note: to keep the current resolver, specify `workspace.resolver = "1"` in the workspace root's manifest
note: to use the edition 2021 resolver, specify `workspace.resolver = "2"` in the workspace root's manifest
```

* Fix clippy warnings in newer Rust compilers.

* Upgrade to bitflags 2

Resolves clippy errors that are now caught: https://github.com/bitflags/bitflags/pull/373

* Add test highlighting the problem

* Fix background thread pool unbound placement

Any threads we spawn inherit our affinity. This means that background
thread pools for any executors that don't have an affinity of "Unbound"
will only ever run on the executor's CPU rather than distributing work
across many CPUs. This seems undesirable given that the background
thread pool placement can be set independent of the executor it's
associated with. Additionally, the background thread pool placement by
default uses the pool placement so this has no impact on background
threads that didn't have a custom placement set (& if the placement was
set, it was likely the wrong thing happening).