Make it more straightforward to send files across threads safely (#619)
* Expose the inode / dev major / dev minor for the opened file

* Make it more straightforward to send files across threads safely

Introduce an OwnedDmaFile struct that a DmaFile can be converted into.
This ensures that the reactor reference that's obtained is always the
one that the DmaFile is running on. This is useful as a way to dupe a
file descriptor, send it to a background thread for some kind of batch
processing. Similarly, the batch processing may create a new file that
contains the result of the batch processing.