don't test for file size on resize_dma_buffer
We already test that the write is successful, and we already test that
we read what we write.

In my local machine on an ext4 filesystem, the file size never changes
after we write this. This is a metadata operation and the filesystem is
free to do this optimization AFAIR, which it likely does, especially
since this is a tmpfile that was not closed. We'd have to have a named
file and fsync() the directory to make sure the metadata always gets
updated, but I don't see the value in testing that.