Fix first sleep taking at least 30-50ms
Noticed this problem in unit tests where the very first timer sleep
would take a very long time because the membarrier was being registered
and tests have more than 1 thread.

Initializing the membarrier strategy before the BlockingThreadPool is
constructed seems like a good idea because it tries to elide the
registration cost if we haven't created any other threads yet. Even if
there are threads, the cost is front-loaded eagerly so it's part of the
cost of creating the executor rather than appearing as a random delay
going to sleep on the io_uring the first time (assuming it could
otherwise wake before the 30-80ms cost observed).

I believe the cost of registration got worse sometime after Linux 6.6
because tests in my project that do something similar started regularly
taking >30ms after upgrading to 6.8 whereas before they were mostly
succeeding < 30ms (it was my grace window for how long a 10ms sleep
could take).

With this change we see that the very first timer now completes within
11ms (I added a grace window of an extra ms in case of CI).