Allow a ReadResult to be converted to a DmaBuffer
This is a hack but it let's us write reads from files without creating
an interim copy. The reason it's a hack is that it's weird to have a
DmaBuffer, which represents a mutable view of data, be convertible from
a ReadResult - notice the panic in attempts to access it mutably.