Fix stat to use the fd instead of the path (#649)
* chore(deps): Upgrade to flume 0.11

Has a race condition fix for async APIs. Not sure if this actually
impacts us but sounds like a worthwhile upgrade.

* Add failing test cases (#648)

Attempting to stat on an unlinked file will error with NotFound when it
shouldn't since we still have the fd. Similarly, invoking stat on a
tempfile will stat the parent directory instead of the file itself.

* Fix stat() behavior to use the fd instead of the path

This fixes the broken test cases - unlinked files will still stat
successfully and temporary files will return stat for the file rather
than the parent directory.

There's probably a negligible performance boost since we don't need to
do anything with the path (I believe it saves a memory allocation + some
other cycles) + the kernel might have to do less work handling the
syscall. Of course, stat should never be in the hot path so it's largely
an academic difference.